requirements.txt
aiofiles==24.1.0
alembic==1.9.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
cryptography==44.0.2
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.1
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.16.1
Flask==2.2.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.19.1
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzbar==0.1.9
requests==2.32.4
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.46
starlette==0.46.1
tomli==2.2.1
typer==0.15.2
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==2.2.2
WTForms==3.1.1
zipp==3.17.0
